+++ packages/package_info_plus/package_info_plus_windows/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.0.0
+
+ - Update a dependency to the latest release.
+
 ## 2.1.0
 
 - Allow win32 3.x to be used.